refactor(cli): extract install-download and install-remove from installer
