+++ packages/xinity-cli/src/commands/update.ts
@@ -8,7 +8,7 @@ import pc from "picocolors";
 import { version } from "../../../../package.json";
 const CLI_VERSION = `v${version}`;
 import { fetchRelease, pickReleaseAsset, type Release } from "../lib/github.ts";
-import { downloadAndVerify, extractCommandArgv } from "../lib/installer.ts";
+import { downloadAndVerify, extractCommandArgv } from "../lib/install-download.ts";
 import { pass, fail } from "../lib/output.ts";
 import { createLocalHost } from "../lib/host.ts";
